TUC-92 Modified the Hardcoded operations to add their required inputs
This is done in the constructor now.

	modified:   src/tucucore/drugfilevalidator.cpp
	modified:   src/tucucore/hardcodedoperation.cpp
	modified:   src/tucucore/hardcodedoperation.h

+++ src/tucucore/hardcodedoperation.cpp
@@ -117,7 +117,7 @@ OperationEGFRCockcroftGaultGeneral::compute(const OperationInputList &_inputs, d
 
     isMale = (sex > 0.5);
 
-    _result = (140 - age) * weight / creatinine * (isMale ? 1.23 : 1.04);
+    _result = static_cast<double>(140 - age) * weight / creatinine * (isMale ? 1.23 : 1.04);
 
     return true;
 }
